Add support for PostgreSQL backend database
The production version of the code will run on the existing PostgreSQL
database, so add support for it. This would have been easier if Python
could pick a single parameter marking style across all backends, and
if SQLite3 could cope with some more data types (like, proper boolean
support for a start).