Special case vote #3
Vote #3 was 3 separate yes/no/abstain questions and as such was special
cased in the original SPI members site as well. We don't need to vote in
it as its long passed, so just special case it in the output template.
If it is necessary to run a multiple unrelated question vote in the
future this should be done as separate votes - the system supports
multiple votes at once.

Additionally, only output the Condorcet information when a vote is
currently active (it doesn't make sense for #3 at all and is unnecessary
once a vote is completed).