Add display of last active date and a way to update it
As part of keeping the contributing member list to those who are active
we need a way to track them. 2009-11-04.jmd.1 suggests one way of doing
this is if they've voted at the annual board election. We update the
last active date when a member votes, so add a display of this date
and a suitable URL that a nudge email can include if someone is still
active but did not vote.