Add support for adding new votes
Allow members with the createvote flag set to create new votes, and
edit and votes they may have created which have not yet run. Also add
display of a users votes to the main status page, with appropriate links
to edit / vote / see the result depending on if it's still to run, is
active or is over.