Update a member's lastactive info when they become a contributing member
When an AM makes a member a contributing member update their lastactive
field to the current date. We could say that it's when they submitted
the application, but that is unfair if they apply before an election and
are approved after an election - they'd get a ping about not voting
despite being unable to.