Add clean-contrib script for cleaning up inactive contributing members
As per 2009-11-04.jmd.1 add a script to send out notification emails
that a member might be considered in active, and also to flip all such
members to non-contributing status.

First pass; only looks for members with no lastactive date set. Future
work will pull the date of the most recent election and use that as the

cut off.