Enable voting system selection and default to more standard Condorcet
Extend the vote_election table to allow each vote to use a different
voting system, enabling the use of new systems while ensuring old votes
remain fully visible in the system. Use this to enable (and default to)
the more standard Condorcet system which places all unranked votes below
those ranked.

This lays the groundwork for a move to a Single Transferable Vote based
system, which is expected to be done before the 2017 board election.